+++ src/pages/ProfilePage/ProfilePage.jsx
@@ -1,0 +1,10 @@
+import ListCards from "../../components/ListCard/listcards";
+function ProfilePage() {
+	return (
+		<div>
+			<ListCards />
+		</div>
+	);
+}
+
+export default ProfilePage;
